+++ WritingReading/1writeToFile.go
@@ -1,27 +1,27 @@
 package main
 
 import (
-    "os"
-    "log"
+	"log"
+	"os"
 )
 
 func main() {
-    // Open a new file for writing only
-    file, err := os.OpenFile(
-        "leking.txt",
-        os.O_WRONLY|os.O_TRUNC|os.O_CREATE,
-        0666,
-    )
-    if err != nil {
-        log.Fatal(err)
-    }
-    defer file.Close()
+	// Åpner en fil for skriving
+	file, err := os.OpenFile(
+		"leking.txt",
+		os.O_WRONLY|os.O_TRUNC|os.O_CREATE,
+		0666,
+	)
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer file.Close()
 
-    // Write bytes to file
-    byteSlice := []byte("Jeg spiste pizza i dag\n")
-    bytesWritten, err := file.Write(byteSlice)
-    if err != nil {
-        log.Fatal(err)
-    }
-    log.Printf("Wrote %d bytes.\n", bytesWritten)
+	// Skriver bytes til filen
+	byteSlice := []byte("Jeg spiste pizza i dag\n")
+	bytesWritten, err := file.Write(byteSlice)
+	if err != nil {
+		log.Fatal(err)
+	}
+	log.Printf("Wrote %d bytes.\n", bytesWritten)
 }
